@@ -1,36 +1,30 @@
 {
-  "name": "zacremboldt.netlify.app",
+  "name": "personal",
+  "version": "1.0.0",
   "private": true,
-  "description": "The personal landing page of Zac Remboldt – Designer // Developer",
-  "version": "0.1.0",
-  "author": "Zac Remboldt <zremboldt@gmail.com>",
-  "dependencies": {
-    "gatsby": "^2.1.4",
-    "gatsby-plugin-manifest": "^2.0.17",
-    "gatsby-plugin-offline": "^2.0.24",
-    "gatsby-plugin-react-helmet": "^3.0.6",
-    "gatsby-plugin-sass": "^2.0.10",
-    "node-sass": "^4.11.0",
-    "prop-types": "^15.6.2",
-    "react": "^16.8.2",
-    "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^16.7.0",
-    "react-helmet": "^5.2.0",
-    "react-spring": "^8.0.9"
-  },
+  "description": "personal",
+  "author": "Zac Remboldt",
   "keywords": [
     "gatsby"
   ],
-  "license": "MIT",
   "scripts": {
-    "build": "gatsby build",
     "develop": "gatsby develop",
-    "start": "npm run develop",
+    "start": "gatsby develop",
+    "build": "gatsby build",
     "serve": "gatsby serve",
-    "test": "echo \"Write tests! -> https://gatsby.app/unit-testing\""
+    "clean": "gatsby clean"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/zremboldt/personal-site"
+  "dependencies": {
+    "gatsby": "^3.4.1",
+    "gatsby-plugin-manifest": "^3.5.0",
+    "gatsby-plugin-react-helmet": "^4.5.0",
+    "gatsby-plugin-sass": "^4.5.0",
+    "gatsby-plugin-sitemap": "^4.1.0",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-helmet": "^6.1.0",
+    "react-spring": "^9.1.2",
+    "sass": "^1.32.13",
+    "use-clipboard-copy": "^0.2.0"
   }
 }
