@@ -1,39 +1,26 @@
-import React, { Component } from "react";
-import ArrowDown from "../images/ArrowDown";
-import Quote from "../images/Quote";
+import React, { useState } from "react";
+import ArrowDown from "../images/arrow-down";
+import Quote from "../images/quote";
 
-export default class Strength extends Component {
-  state = {
-    open: false
-  };
+export default function Strength({ strength, description }) {
+  const [isVisible, setIsVisible] = useState(false)
 
-  handleMouseEnter = () => {
-    this.setState({ open: true });
-  };
-  handleMouseLeave = () => {
-    this.setState({ open: false });
-  };
-
-  render() {
-    const { open } = this.state;
-    const { strength, description } = this.props.strength;
-    return (
-      <li
-        className="strength"
-        onMouseEnter={this.handleMouseEnter}
-        onMouseLeave={this.handleMouseLeave}
-      >
-        <div className="strengthLabel">
-          <p>{strength}</p>
-          <ArrowDown />
+  return (
+    <li
+      className="strength"
+      onMouseEnter={() => setIsVisible(true)}
+      onMouseLeave={() => setIsVisible(false)}
+    >
+      <div className="strengthLabel">
+        <p>{strength}</p>
+        <ArrowDown />
+      </div>
+      {isVisible && (
+        <div className="dropdown">
+          <Quote />
+          <p>{description}</p>
         </div>
-        {open && (
-          <div className="dropdown">
-            <Quote />
-            <p>{description}</p>
-          </div>
-        )}
-      </li>
-    );
-  }
+      )}
+    </li>
+  );
 }
